LPS-197299 Rename
As per https://github.com/liferay-frontend/liferay-portal/pull/3703#discussion_r1341388941
