+++ modules/apps/frontend-data-set/frontend-data-set-web/types/src/main/resources/META-INF/resources/management_bar/components/filters/implementation/SelectionFilter.d.ts
